refactor: component라는 postfix 제거
