+++ frontend/components/kanban/card/CardApplicantStatusLabel.tsx
